feat: Add "tour" section to navigation and home page
The code changes include adding the "tour" section to the navigation component and the home page. This allows users to easily navigate to the tour section of the website. The sections array in the BackToTop component has also been updated to include the "tour" section.

Recent user commits:
- Removed comment
- Added comment to music videos

Recent repository commits:
- Merge pull request #9 from daniel-cox/Matthew: TODO adjust colors and set theme of the site
- TODO adjust colors and set theme of the site
- Merge pull request #8 from daniel-cox/Matthew: Matthew
- Updated logo color array
- Merge branch 'development' of https://github.com/daniel-cox/Jawfane into Matthew
- Added placeholder prices to merch
- Added sample images for merch
- Updated layout, implemented mobile changes